+++ sciencebeam_gym/preprocess/annotation/find_line_numbers_test.py
@@ -8,7 +8,7 @@ from sciencebeam_gym.structured_document import (
   SimpleToken
 )
 
-from sciencebeam_gym.preprocess.find_line_number import (
+from sciencebeam_gym.preprocess.annotation.find_line_number import (
   find_line_number_tokens
 )
