+++ sciencebeam_gym/preprocess/preprocessing_utils.py
@@ -3,7 +3,6 @@ from __future__ import absolute_import
 import os
 import logging
 from io import BytesIO
-from zipfile import ZipFile, ZIP_DEFLATED
 from functools import reduce # pylint: disable=W0622
 
 from six import iteritems
@@ -43,7 +42,7 @@ from sciencebeam_gym.structured_document.svg import (
   SvgStructuredDocument
 )
 
-from sciencebeam_gym.preprocess.annotator import (
+from sciencebeam_gym.preprocess.annotation.annotator import (
   Annotator,
   DEFAULT_ANNOTATORS
 )
@@ -52,7 +51,7 @@ from sciencebeam_gym.alignment.align import (
   native_enabled as align_native_enabled
 )
 
-from sciencebeam_gym.preprocess.matching_annotator import (
+from sciencebeam_gym.preprocess.annotation.matching_annotator import (
   MatchingAnnotator,
   xml_root_to_target_annotations
 )
